[feature] AgileScreen: Agile rate compact screen done
